Refactor model startup
